Finished helper function in JsonUtils to parse json into Sandwich Object.
Added helper functions to Sandwich object to get List items (alsoKnownAs and ingredients) returned as multiline string, in the format we would expect to display them. Done here to simplify if that data gets displayed somewhere else later.

Finished populateUI function in DetailActivity, passing it the sandwich object. This data is applied to the TextViews.

Changed activity_detail to have a ScrollView with ConstaintLayout instead of FrameLayout.